+++ component/pom.xml
@@ -89,6 +89,11 @@
 <!--            <version>2.8.1</version>-->
 <!--        </dependency>-->
         <dependency>
+            <groupId>org.conscrypt</groupId>
+            <artifactId>conscrypt-openjdk-uber</artifactId>
+            <version>2.5.2</version>
+        </dependency>
+        <dependency>
             <groupId>org.wso2.carbon.metrics</groupId>
             <artifactId>org.wso2.carbon.metrics.core</artifactId>
             <version>2.3.15</version>
